+++ src/component/domain/main-banner/MainBanner.ts
@@ -1,5 +1,5 @@
-import { MovieData } from '../../../types/movie';
-import Button from '../button/Button';
+import { MovieData } from '../../../../types/movie';
+import Button from '../../common/button/Button';
 
 interface MovieBannerProps {
   data: MovieData;
@@ -24,11 +24,14 @@ class MainBanner {
        </div>
        
        <div class="main-banner__info">
+
          <div class="main-banner__rate">
            <img src="./star_empty.png" class="main-banner__rating-star" />
            <span class="main-banner__rate-value text-subtitle">${this.#data.score}</span>
          </div>
+
          <div class="main-banner__title text-title">${this.#data.title}</div>
+
          <div class="main-banner__button">${this.#detailButtonElement()}</div>
        </div>
     `;
