+++ src/component/domain/main-banner/main-banner.css
