+++ src/component/domain/movie-grid/movie-grid.css
