+++ src/component/domain/movie-item/movie-item.css
