+++ src/pages/search-page/SearchPage.ts
@@ -1,6 +1,6 @@
-import Button from '../../component/button/Button';
-import MovieGrid from '../../component/movie-grid/MovieGrid';
-import { Title } from '../../component/title/Title';
+import Button from '../../component/common/button/Button';
+import MovieGrid from '../../component/domain/movie-grid/MovieGrid';
+import { Title } from '../../component/common/title/Title';
 import { SYSTEM_CONSTANTS } from '../../constants/systemConstants';
 import { extractedeData } from '../../domain/APIManager';
 import searchPageLoadingTemplate from './loadingTemplate';
